add projects table to DB

+++ Portfolio/Models/Project.cs
@@ -5,7 +5,7 @@ namespace Portfolio.Models
     public class Project
     {
 
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
         public string Name { get; set; }
